paint_lib: simplify handlers and fix warnings

+++ 3rdlib/paint_lib/qtmypaint/mphandler.h
@@ -61,7 +61,7 @@ public:
 
     void startStroke();
     void strokeTo(float x, float y, float pressure, float xtilt, float ytilt, double dtime);
-    void strokeTo(double x, double y);
+    void strokeTo(float x, float y);
     void endStroke();
 
     float getBrushValue(MyPaintBrushSetting setting);
